start relocating code to generate notifiction in its own dedicated source file
